Introduce BsonWriterAdapter superclass
JAVA-3328
